Clean up runSoot usages
